+++ richat/src/source.rs
@@ -1,12 +1,23 @@
 use {
     crate::{
         channel::ParsedMessage,
-        config::{ConfigChannelSource, ConfigChannelSourceShared, ConfigGrpcClientSource},
+        config::{
+            ConfigChannelSource, ConfigChannelSourceGeneral, ConfigChannelSourceReconnect,
+            ConfigGrpcClientSource,
+        },
     },
-    futures::stream::{BoxStream, StreamExt},
+    futures::{
+        ready,
+        stream::{try_unfold, BoxStream, Stream, StreamExt},
+    },
     maplit::hashmap,
-    richat_client::{error::ReceiveError, grpc::GrpcClientBuilderError, quic::QuicConnectError},
-    richat_filter::message::{Message, MessageParseError},
+    pin_project_lite::pin_project,
+    richat_client::{
+        grpc::{ConfigGrpcClient, GrpcClientBuilderError},
+        quic::{ConfigQuicClient, QuicConnectError},
+        tcp::ConfigTcpClient,
+    },
+    richat_filter::message::{Message, MessageParseError, MessageParserEncoding},
     richat_proto::{
         geyser::{
             CommitmentLevel as CommitmentLevelProto, SubscribeRequest,
@@ -16,8 +27,15 @@ use {
         },
         richat::{GrpcSubscribeRequest, RichatFilter},
     },
-    std::{collections::HashMap, io},
+    std::{
+        collections::HashMap,
+        fmt, io,
+        pin::Pin,
+        task::{Context, Poll},
+    },
     thiserror::Error,
+    tokio::time::{sleep, Duration},
+    tracing::{error, info},
 };
 
 #[derive(Debug, Error)]
@@ -31,7 +49,7 @@ enum ConnectError {
 }
 
 #[derive(Debug, Error)]
-enum SubscribePlainError {
+enum SubscribeError {
     #[error(transparent)]
     Connect(#[from] ConnectError),
     #[error(transparent)]
@@ -40,99 +58,245 @@ enum SubscribePlainError {
     SubscribeGrpc(#[from] tonic::Status),
 }
 
-async fn subscribe_plain(
-    config: ConfigChannelSource,
-) -> Result<
-    (
-        ConfigChannelSourceShared,
-        BoxStream<'static, Result<Vec<u8>, ReceiveError>>,
-    ),
-    SubscribePlainError,
-> {
-    let create_filter = |disable_accounts: bool| {
+#[derive(Debug, Error)]
+pub enum ReceiveError {
+    #[error(transparent)]
+    Receive(#[from] richat_client::error::ReceiveError),
+    #[error(transparent)]
+    Parse(#[from] MessageParseError),
+}
+
+#[derive(Debug, Clone)]
+enum SubscriptionConfig {
+    Quic {
+        config: ConfigQuicClient,
+    },
+    Tcp {
+        config: ConfigTcpClient,
+    },
+    Grpc {
+        source: ConfigGrpcClientSource,
+        config: ConfigGrpcClient,
+    },
+}
+
+impl SubscriptionConfig {
+    fn new(config: ConfigChannelSource) -> (Self, ConfigChannelSourceGeneral) {
+        match config {
+            ConfigChannelSource::Quic { general, config } => (Self::Quic { config }, general),
+            ConfigChannelSource::Tcp { general, config } => (Self::Tcp { config }, general),
+            ConfigChannelSource::Grpc {
+                general,
+                source,
+                config,
+            } => (Self::Grpc { source, config }, general),
+        }
+    }
+}
+
+pin_project! {
+    struct Subscription {
+        stream: BoxStream<'static, Result<Vec<u8>, richat_client::error::ReceiveError>>,
+        parser: MessageParserEncoding,
+        index: usize,
+    }
+}
+
+impl fmt::Debug for Subscription {
+    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
+        f.debug_struct("Subscription").finish()
+    }
+}
+
+impl Stream for Subscription {
+    type Item = Result<(usize, ParsedMessage), ReceiveError>;
+
+    fn poll_next(mut self: Pin<&mut Self>, cx: &mut Context<'_>) -> Poll<Option<Self::Item>> {
+        loop {
+            let value = ready!(self.stream.poll_next_unpin(cx));
+            return Poll::Ready(match value {
+                Some(Ok(data)) => match Message::parse(data, self.parser) {
+                    Ok(message) => Some(Ok((self.index, message.into()))),
+                    Err(MessageParseError::InvalidUpdateMessage("Ping")) => continue,
+                    Err(error) => Some(Err(error.into())),
+                },
+                Some(Err(error)) => Some(Err(error.into())),
+                None => None,
+            });
+        }
+    }
+}
+
+impl Subscription {
+    async fn new(
+        config: SubscriptionConfig,
+        disable_accounts: bool,
+        parser: MessageParserEncoding,
+        index: usize,
+    ) -> Result<Self, SubscribeError> {
+        let stream = match config {
+            SubscriptionConfig::Quic { config } => {
+                let connection = config.connect().await.map_err(ConnectError::Quic)?;
+                let filter = Self::create_richat_filter(disable_accounts);
+                connection.subscribe(None, filter).await?.boxed()
+            }
+            SubscriptionConfig::Tcp { config } => {
+                let connection = config.connect().await.map_err(ConnectError::Tcp)?;
+                let filter = Self::create_richat_filter(disable_accounts);
+                connection.subscribe(None, filter).await?.boxed()
+            }
+            SubscriptionConfig::Grpc { source, config } => {
+                let mut connection = config.connect().await.map_err(ConnectError::Grpc)?;
+                match source {
+                    ConfigGrpcClientSource::DragonsMouth => connection
+                        .subscribe_dragons_mouth_once(Self::create_dragons_mouth_filter(
+                            disable_accounts,
+                        ))
+                        .await?
+                        .boxed(),
+                    ConfigGrpcClientSource::Richat => connection
+                        .subscribe_richat(GrpcSubscribeRequest {
+                            replay_from_slot: None,
+                            filter: Self::create_richat_filter(disable_accounts),
+                        })
+                        .await?
+                        .boxed(),
+                }
+            }
+        };
+
+        Ok(Self {
+            stream,
+            parser,
+            index,
+        })
+    }
+
+    const fn create_richat_filter(disable_accounts: bool) -> Option<RichatFilter> {
         Some(RichatFilter {
             disable_accounts,
             disable_transactions: false,
             disable_entries: false,
         })
-    };
+    }
 
-    Ok(match config {
-        ConfigChannelSource::Quic { general, config } => {
-            let connection = config.connect().await.map_err(ConnectError::Quic)?;
-            let filter = create_filter(general.disable_accounts);
-            (general, connection.subscribe(None, filter).await?.boxed())
+    fn create_dragons_mouth_filter(disable_accounts: bool) -> SubscribeRequest {
+        SubscribeRequest {
+            accounts: if disable_accounts {
+                HashMap::new()
+            } else {
+                hashmap! { "".to_owned() => SubscribeRequestFilterAccounts::default() }
+            },
+            slots: hashmap! { "".to_owned() => SubscribeRequestFilterSlots {
+                filter_by_commitment: Some(false),
+                interslot_updates: Some(true),
+            } },
+            transactions: hashmap! { "".to_owned() => SubscribeRequestFilterTransactions::default() },
+            transactions_status: HashMap::new(),
+            blocks: HashMap::new(),
+            blocks_meta: hashmap! { "".to_owned() => SubscribeRequestFilterBlocksMeta::default() },
+            entry: hashmap! { "".to_owned() => SubscribeRequestFilterEntry::default() },
+            commitment: Some(CommitmentLevelProto::Processed as i32),
+            accounts_data_slice: vec![],
+            ping: None,
+            from_slot: None,
         }
-        ConfigChannelSource::Tcp { general, config } => {
-            let connection = config.connect().await.map_err(ConnectError::Tcp)?;
-            let filter = create_filter(general.disable_accounts);
-            (general, connection.subscribe(None, filter).await?.boxed())
-        }
-        ConfigChannelSource::Grpc {
-            general,
-            source,
-            config,
-        } => {
-            let mut connection = config.connect().await.map_err(ConnectError::Grpc)?;
-            let stream = match source {
-                ConfigGrpcClientSource::DragonsMouth => connection
-                    .subscribe_dragons_mouth_once(SubscribeRequest {
-                        accounts: if general.disable_accounts {
-                            HashMap::new()
-                        } else {
-                            hashmap! { "".to_owned() => SubscribeRequestFilterAccounts::default() }
-                        },
-                        slots: hashmap! { "".to_owned() => SubscribeRequestFilterSlots {
-                            filter_by_commitment: Some(false),
-                            interslot_updates: Some(true),
-                        } },
-                        transactions: hashmap! { "".to_owned() => SubscribeRequestFilterTransactions::default() },
-                        transactions_status: HashMap::new(),
-                        blocks: HashMap::new(),
-                        blocks_meta: hashmap! { "".to_owned() => SubscribeRequestFilterBlocksMeta::default() },
-                        entry: hashmap! { "".to_owned() => SubscribeRequestFilterEntry::default() },
-                        commitment: Some(CommitmentLevelProto::Processed as i32),
-                        accounts_data_slice: vec![],
-                        ping: None,
-                        from_slot: None,
-                    })
-                    .await?
-                    .boxed(),
-                ConfigGrpcClientSource::Richat => connection
-                    .subscribe_richat(GrpcSubscribeRequest {
-                        replay_from_slot: None,
-                        filter: create_filter(general.disable_accounts),
-                    })
-                    .await?
-                    .boxed(),
-            };
-            (general, stream)
-        }
-    })
+    }
 }
 
-#[derive(Debug, Error)]
-pub enum SubscribeError {
-    #[error(transparent)]
-    Receive(#[from] ReceiveError),
-    #[error(transparent)]
-    Parse(#[from] MessageParseError),
+#[derive(Debug)]
+struct Backoff {
+    current_interval: Duration,
+    initial_interval: Duration,
+    max_interval: Duration,
+    multiplier: f64,
+}
+
+impl Backoff {
+    const fn new(config: ConfigChannelSourceReconnect) -> Self {
+        Self {
+            current_interval: config.initial_interval,
+            initial_interval: config.initial_interval,
+            max_interval: config.max_interval,
+            multiplier: config.multiplier,
+        }
+    }
+
+    async fn sleep(&mut self) {
+        sleep(self.current_interval).await;
+        self.current_interval = self
+            .current_interval
+            .mul_f64(self.multiplier)
+            .min(self.max_interval);
+    }
+
+    fn reset(&mut self) {
+        self.current_interval = self.initial_interval;
+    }
 }
 
 pub async fn subscribe(
     config: ConfigChannelSource,
-) -> anyhow::Result<BoxStream<'static, Result<ParsedMessage, SubscribeError>>> {
-    let (config, stream) = subscribe_plain(config).await?;
-    Ok(stream
-        .filter_map(move |value| async move {
-            match value {
-                Ok(data) => match Message::parse(data, config.parser) {
-                    Ok(message) => Some(Ok(message.into())),
-                    Err(MessageParseError::InvalidUpdateMessage("Ping")) => None,
-                    Err(error) => Some(Err(error.into())),
-                },
-                Err(error) => Some(Err(error.into())),
+    index: usize,
+) -> anyhow::Result<BoxStream<'static, Result<(usize, ParsedMessage), ReceiveError>>> {
+    let (subscription_config, mut config) = SubscriptionConfig::new(config);
+
+    let Some(reconnect) = config.reconnect.take() else {
+        let stream = Subscription::new(
+            subscription_config,
+            config.disable_accounts,
+            config.parser,
+            index,
+        )
+        .await?;
+        info!(name = config.name, "connected");
+        return Ok(stream.boxed());
+    };
+
+    let backoff = Backoff::new(reconnect);
+    let stream = try_unfold(
+        (backoff, subscription_config, config, None),
+        move |mut state: (
+            Backoff,
+            SubscriptionConfig,
+            ConfigChannelSourceGeneral,
+            Option<Subscription>,
+        )| async move {
+            loop {
+                if let Some(stream) = state.3.as_mut() {
+                    match stream.next().await {
+                        Some(Ok(message)) => return Ok(Some((message, state))),
+                        Some(Err(error)) => {
+                            error!(name = state.2.name, ?error, "failed to receive")
+                        }
+                        None => {
+                            error!(name = state.2.name, "stream is finished")
+                        }
+                    }
+                    state.3 = None;
+                    state.0.sleep().await;
+                } else {
+                    match Subscription::new(
+                        state.1.clone(),
+                        state.2.disable_accounts,
+                        state.2.parser,
+                        index,
+                    )
+                    .await
+                    {
+                        Ok(stream) => {
+                            info!(name = state.2.name, "connected");
+                            state.3 = Some(stream);
+                            state.0.reset();
+                        }
+                        Err(error) => {
+                            error!(name = state.2.name, ?error, "failed to connect");
+                            state.0.sleep().await;
+                        }
+                    }
+                }
             }
-        })
-        .boxed())
+        },
+    );
+    Ok(stream.boxed())
 }